# Optimize large file uploads