@christek91 2018-12-04
Aside from testing the performance improvements of this code, I've also manually verified that the MD5 checksums of files uploaded with the new code still match their originals, so I believe that uploading still works correctly.