@christek91 2019-01-07
@minrk do you know how we could get this packaged into a minor version release? My team would like to use this change but would also like to avoid forking jupyter/notebook internally if possible.